Revert "Revert "#98 move query parser""
This reverts commit 35498a4c9d63c356017e001aae534bb534ef60f5.
