+++ metadata-engine/src/main/resources/META-INF/spring.factories
@@ -6,5 +6,4 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   io.github.sergeivisotsky.metadata.engine.config.ChartMetadataConfig,\
   io.github.sergeivisotsky.metadata.engine.config.PropertiesConfig,\
   io.github.sergeivisotsky.metadata.engine.config.CacheableMetadataConfig,\
-  io.github.sergeivisotsky.metadata.engine.filtering.config.FilteringConfig,\
   io.github.sergeivisotsky.metadata.engine.jdbc.sqlparser.config.SQLParserConfig
